Some refatoring of the ScheduleManager
   * Make the RemoteCacheProducer to focus on counter function
   * Remote Abstract CacheProducer
