+++ src/main/java/org/commonjava/indy/service/scheduler/data/ispn/ISPNScheduleManager.java
@@ -20,10 +20,10 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import io.quarkus.runtime.Startup;
 import org.commonjava.indy.service.scheduler.data.ScheduleManager;
 import org.commonjava.indy.service.scheduler.data.StandaloneScheduleManager;
+import org.commonjava.indy.service.scheduler.data.ispn.local.CacheHandle;
 import org.commonjava.indy.service.scheduler.event.ScheduleTriggerEvent;
 import org.commonjava.indy.service.scheduler.event.kafka.KafkaEventUtils;
 import org.commonjava.indy.service.scheduler.exception.SchedulerException;
-import org.commonjava.indy.service.scheduler.model.ContentExpiration;
 import org.commonjava.indy.service.scheduler.model.Expiration;
 import org.commonjava.indy.service.scheduler.model.ExpirationSet;
 import org.commonjava.indy.service.scheduler.model.ScheduleKey;
