--- src/main/java/org/commonjava/indy/service/scheduler/data/ispn/RemoteCacheHandle.java
@@ -1,44 +1,0 @@
-/**
- * Copyright (C) 2011-2020 Red Hat, Inc. (https://github.com/Commonjava/indy)
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *         http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.commonjava.indy.service.scheduler.data.ispn;
-
-import org.commonjava.indy.service.scheduler.data.metrics.DefaultMetricsManager;
-import org.infinispan.client.hotrod.RemoteCache;
-
-import static org.commonjava.indy.service.scheduler.data.metrics.NameUtils.name;
-
-public class RemoteCacheHandle<K, V>
-        extends BasicCacheHandle<K, V>
-{
-
-    public RemoteCacheHandle( String named, RemoteCache<K, V> cache, DefaultMetricsManager metricsManager,
-                              String metricPrefix )
-    {
-        super( named, cache, metricsManager, metricPrefix );
-    }
-
-    public RemoteCacheHandle( String named, RemoteCache<K, V> cache )
-    {
-        this( named, cache, null, null );
-    }
-
-    @Override
-    protected String getMetricName( String opName )
-    {
-        return name( getMetricPrefix(), cache.getName(), "remote", opName );
-    }
-
-}
